Create Student class
I added 2 fields to Student: name and image. The name is currently an int field but idk why its supposed to. I think the strings resource is where we are supposed to have all the names and each name has an id so we have to use that integer as the name.